requirements.txt
fastapi==0.115.0
uvicorn==0.33.0
sqlalchemy==2.0.36
pydantic==2.10.6
python-multipart==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.1.1
bcrypt==4.1.2
